Minor tweaks to auto-fiod to make it more useful in creating a random fiod configuration file.
Fixed a long standing issue with fiod where it couldn't display the
configuration data for each job until the target file(s) were filled.
Pretty dumb, just split out the filling code from the initialization so
they could be called separately. Now, the jobs are initialized and the
configuration data is displayed if requested. Then the file(s) are filled.

Only bother to display progress in percentage when the number of
jobs is less than 5. Since the code pretty prints the filling data on
a single line more than 5 files will go over 80 characters and the display
become garbage so it's pointless to track at that detail.