Major clean up of the main loop. Reduced the code to use the tracker class when filling, running, and final clean up.
Stat output is handle within the tracker class instead of having a
timer channel in the stat code doing the display. Now stat has a string
method to return what's needed.

Corrected the interrupt code. Previously it would only work when in
the main loop. Now the interrupt is handled by the tracker class so
everything can be interrupted.